pyproject.toml
[project]
name = "ddmy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyside6-fluent-widgets[full]>=1.7.1",
]
